+++ src/main/kotlin/com/chromaway/postchain/base/data/BaseBlockchainConfiguration.kt
@@ -63,7 +63,7 @@ open class BaseBlockchainConfiguration(override val chainID: Long, val config: C
     override fun makeBlockQueries(storage: Storage): BlockQueries {
         val blockSigningPrivateKey = config.getString("blocksigningprivkey").hexStringToByteArray()
         val blockSigningPublicKey = secp256k1_derivePubKey(blockSigningPrivateKey)
-        return BaseBlockQueries(this, storage, blockStore, chainID.toInt(), blockSigningPublicKey)
+        return BaseBlockQueries(this, storage, blockStore, chainID, blockSigningPublicKey)
     }
 
     override fun initializeDB(ctx: EContext) {
